更改舵机初始化函数的参数类型为u16,添加底盘PID

+++ DEVICE/inc/time1.h
@@ -3,7 +3,7 @@
 
 #include "stm32f10x.h"
 
-void Time1_Pwm_Init(uint32_t arr, uint32_t psc);
+void Time1_Pwm_Init(u16 arr, u16 psc);
 void Time1_Pwm_Init1(u16 arr, u16 psc);
 
 #endif //_TIME1_H_
